Move Admin samples into proper package
